🎨 Add waitlist and refactor responses
Also admin view is its own component now.
